I said not to do this
this is hardcore evil spaghetti code if gate spam type shit.

if you find yourself using three or more if gates in a row, there is always a better way to do it!

spaghetti code aside, if gate spam *will* cause performance issues. It also means a hell of a lot more work for you!

Work smarter, not harder!